+++ src/com/qdacity/server/project/CodeSystemEndpoint.java
@@ -1,11 +1,15 @@
 package com.qdacity.server.project;
 
+import com.qdacity.Constants;
+import com.qdacity.server.Authorization;
 import com.qdacity.server.PMF;
 import com.google.api.server.spi.config.Api;
 import com.google.api.server.spi.config.ApiMethod;
 import com.google.api.server.spi.config.ApiNamespace;
 import com.google.api.server.spi.response.CollectionResponse;
+import com.google.api.server.spi.response.UnauthorizedException;
 import com.google.appengine.api.datastore.Cursor;
+import com.google.appengine.api.users.User;
 import com.google.appengine.datanucleus.query.JDOCursorHelper;
 
 import java.util.HashMap;
@@ -25,49 +29,57 @@ public class CodeSystemEndpoint {
 	/**
 	 * This method lists all the entities inserted in datastore.
 	 * It uses HTTP GET method and paging support.
+	 * @param user 
 	 *
 	 * @return A CollectionResponse class containing the list of all entities
 	 * persisted and a cursor to the next page.
+	 * @throws UnauthorizedException 
 	 */
 	@SuppressWarnings({ "unchecked", "unused" })
-	@ApiMethod(name = "codesystem.listCodeSystem")
+	@ApiMethod(name = "codesystem.listCodeSystem",  scopes = {Constants.EMAIL_SCOPE},
+			clientIds = {Constants.WEB_CLIENT_ID, 
+		     com.google.api.server.spi.Constant.API_EXPLORER_CLIENT_ID},
+		     audiences = {Constants.WEB_CLIENT_ID})
 	public CollectionResponse<CodeSystem> listCodeSystem(
 			@Nullable @Named("cursor") String cursorString,
-			@Nullable @Named("limit") Integer limit) {
-
-		PersistenceManager mgr = null;
-		Cursor cursor = null;
-		List<CodeSystem> execute = null;
-
-		try {
-			mgr = getPersistenceManager();
-			Query query = mgr.newQuery(CodeSystem.class);
-			if (cursorString != null && cursorString != "") {
-				cursor = Cursor.fromWebSafeString(cursorString);
-				HashMap<String, Object> extensionMap = new HashMap<String, Object>();
-				extensionMap.put(JDOCursorHelper.CURSOR_EXTENSION, cursor);
-				query.setExtensions(extensionMap);
-			}
-
-			if (limit != null) {
-				query.setRange(0, limit);
-			}
-
-			execute = (List<CodeSystem>) query.execute();
-			cursor = JDOCursorHelper.getCursor(execute);
-			if (cursor != null)
-				cursorString = cursor.toWebSafeString();
-
-			// Tight loop for fetching all entities from datastore and accomodate
-			// for lazy fetch.
-			for (CodeSystem obj : execute)
-				;
-		} finally {
-			mgr.close();
-		}
-
-		return CollectionResponse.<CodeSystem> builder().setItems(execute)
-				.setNextPageToken(cursorString).build();
+			@Nullable @Named("limit") Integer limit, User user) throws UnauthorizedException {
+		if (user == null) throw new UnauthorizedException("User is Not Valid");
+		
+		throw new UnauthorizedException("User not authorized"); // TODO currently no user is authorized to list all codesystems
+		
+//		PersistenceManager mgr = null;
+//		Cursor cursor = null;
+//		List<CodeSystem> execute = null;
+//
+//		try {
+//			mgr = getPersistenceManager();
+//			Query query = mgr.newQuery(CodeSystem.class);
+//			if (cursorString != null && cursorString != "") {
+//				cursor = Cursor.fromWebSafeString(cursorString);
+//				HashMap<String, Object> extensionMap = new HashMap<String, Object>();
+//				extensionMap.put(JDOCursorHelper.CURSOR_EXTENSION, cursor);
+//				query.setExtensions(extensionMap);
+//			}
+//
+//			if (limit != null) {
+//				query.setRange(0, limit);
+//			}
+//
+//			execute = (List<CodeSystem>) query.execute();
+//			cursor = JDOCursorHelper.getCursor(execute);
+//			if (cursor != null)
+//				cursorString = cursor.toWebSafeString();
+//
+//			// Tight loop for fetching all entities from datastore and accomodate
+//			// for lazy fetch.
+//			for (CodeSystem obj : execute)
+//				;
+//		} finally {
+//			mgr.close();
+//		}
+//
+//		return CollectionResponse.<CodeSystem> builder().setItems(execute)
+//				.setNextPageToken(cursorString).build();
 	}
 
 	/**
@@ -75,17 +87,29 @@ public class CodeSystemEndpoint {
 	 *
 	 * @param id the primary key of the java bean.
 	 * @return The entity with primary key id.
+	 * @throws UnauthorizedException 
 	 */
-	@ApiMethod(name = "codesystem.getCodeSystem")
+	@ApiMethod(name = "codesystem.getCodeSystem",  scopes = {Constants.EMAIL_SCOPE},
+			clientIds = {Constants.WEB_CLIENT_ID, 
+		     com.google.api.server.spi.Constant.API_EXPLORER_CLIENT_ID},
+		     audiences = {Constants.WEB_CLIENT_ID})
 	public CollectionResponse<Code> getCodeSystem(@Named("id") Long id,
 			@Nullable @Named("cursor") String cursorString,
-			@Nullable @Named("limit") Integer limit) {
+			@Nullable @Named("limit") Integer limit,
+			User user) throws UnauthorizedException {
+		
 		PersistenceManager mgr = null;
 		Cursor cursor = null;
 		List<Code> execute = null;
 
 		try {
 			mgr = getPersistenceManager();
+			
+			// Check if user is Authorized
+			Long projectID = CodeSystemEndpoint.getProjectIdFromCodesystem(id);
+			Authorization.checkAuthorization(projectID, user);
+			// User is authorized
+			
 			Query query = mgr.newQuery(Code.class);
 			if (cursorString != null && cursorString != "") {
 				cursor = Cursor.fromWebSafeString(cursorString);
@@ -120,18 +144,28 @@ public class CodeSystemEndpoint {
 				.setNextPageToken(cursorString).build();
 	}
 
+
 	/**
 	 * This inserts a new entity into App Engine datastore. If the entity already
 	 * exists in the datastore, an exception is thrown.
 	 * It uses HTTP POST method.
 	 *
 	 * @param codesystem the entity to be inserted.
+	 * @param user 
 	 * @return The inserted entity.
+	 * @throws UnauthorizedException 
 	 */
-	@ApiMethod(name = "codesystem.insertCodeSystem")
-	public CodeSystem insertCodeSystem(CodeSystem codesystem) {
+	@ApiMethod(name = "codesystem.insertCodeSystem",  scopes = {Constants.EMAIL_SCOPE},
+			clientIds = {Constants.WEB_CLIENT_ID, 
+		     com.google.api.server.spi.Constant.API_EXPLORER_CLIENT_ID},
+		     audiences = {Constants.WEB_CLIENT_ID})
+	public CodeSystem insertCodeSystem(CodeSystem codesystem, User user) throws UnauthorizedException {
+		//Check if user is authorized
+		Authorization.checkAuthorization(codesystem, user);
+		
 		PersistenceManager mgr = getPersistenceManager();
 		try {
+						
 			if (containsCodeSystem(codesystem)) {
 				throw new EntityExistsException("Object already exists");
 			}
@@ -148,12 +182,21 @@ public class CodeSystemEndpoint {
 	 * It uses HTTP PUT method.
 	 *
 	 * @param codesystem the entity to be updated.
+	 * @param user 
 	 * @return The updated entity.
+	 * @throws UnauthorizedException 
 	 */
-	@ApiMethod(name = "codesystem.updateCodeSystem")
-	public CodeSystem updateCodeSystem(CodeSystem codesystem) {
+	@ApiMethod(name = "codesystem.updateCodeSystem",  scopes = {Constants.EMAIL_SCOPE},
+			clientIds = {Constants.WEB_CLIENT_ID, 
+		     com.google.api.server.spi.Constant.API_EXPLORER_CLIENT_ID},
+		     audiences = {Constants.WEB_CLIENT_ID})
+	public CodeSystem updateCodeSystem(CodeSystem codesystem, User user) throws UnauthorizedException {
 		PersistenceManager mgr = getPersistenceManager();
 		try {
+			// Check if user is Authorized
+			Authorization.checkAuthorization(codesystem, user);
+			// User is authorized
+						
 			if (!containsCodeSystem(codesystem)) {
 				throw new EntityNotFoundException("Object does not exist");
 			}
@@ -169,18 +212,32 @@ public class CodeSystemEndpoint {
 	 * It uses HTTP DELETE method.
 	 *
 	 * @param id the primary key of the entity to be deleted.
+	 * @throws UnauthorizedException 
 	 */
-	@ApiMethod(name = "codesystem.removeCodeSystem")
-	public void removeCodeSystem(@Named("id") Long id) {
+	@ApiMethod(name = "codesystem.removeCodeSystem",  scopes = {Constants.EMAIL_SCOPE},
+			clientIds = {Constants.WEB_CLIENT_ID, 
+		     com.google.api.server.spi.Constant.API_EXPLORER_CLIENT_ID},
+		     audiences = {Constants.WEB_CLIENT_ID})
+	public void removeCodeSystem(@Named("id") Long id, User user) throws UnauthorizedException {
 		PersistenceManager mgr = getPersistenceManager();
 		try {
 			CodeSystem codesystem = mgr.getObjectById(CodeSystem.class, id);
+			
+			// Check if user is authorized
+			Authorization.checkAuthorization(codesystem, user);
+			
 			mgr.deletePersistent(codesystem);
 		} finally {
 			mgr.close();
 		}
 	}
 
+	
+	
+	//
+	// Helper functions
+	//
+	
 	private boolean containsCodeSystem(CodeSystem codesystem) {
 		PersistenceManager mgr = getPersistenceManager();
 		boolean contains = true;
@@ -196,6 +253,29 @@ public class CodeSystemEndpoint {
 
 	private static PersistenceManager getPersistenceManager() {
 		return PMF.get().getPersistenceManager();
+	}
+	
+	public static Long getProjectIdFromCodesystem(Long id) {
+		
+		PersistenceManager mgr = null;
+		Long projectID = (long) -1;
+		try {
+			mgr = getPersistenceManager();
+			
+		Query query = mgr.newQuery(Project.class);
+		
+		query.setFilter( "codesystemID == :theID");
+		Map<String, Long> params = new HashMap();
+		params.put("theID", id);
+		
+		List<Project> projects = (List<Project>) query.executeWithMap(params);
+		Project project = projects.get(0);
+		projectID = project.id;
+		} finally {
+			mgr.close();
+		}
+		
+		return projectID;
 	}
 
 }
